Do depth testing with integer values.
Using floats (and float->ushort->float conversion) introduces errors.
Only GLushort depth buffers work for now...

+++ src/mesa/pipe/softpipe/sp_surface.h
@@ -79,9 +79,9 @@ struct softpipe_surface {
                               GLubyte rgba[NUM_CHANNELS] );
 
    void (*read_quad_z)(struct softpipe_surface *,
-                       GLint x, GLint y, GLfloat zzzz[QUAD_SIZE]);
+                       GLint x, GLint y, GLuint zzzz[QUAD_SIZE]);
    void (*write_quad_z)(struct softpipe_surface *,
-                        GLint x, GLint y, const GLfloat zzzz[QUAD_SIZE]);
+                        GLint x, GLint y, const GLuint zzzz[QUAD_SIZE]);
 };
